:recycle: Moved login, register, and logout handling to index route file
