+++ src/views/login.ejs
@@ -4,7 +4,7 @@
 
 <%- include ('./partials/messages') %>
 
-<form action="/account/login" method="post">
+<form action="/login" method="post">
 	<div>
 		<label for="email">Email</label>
 		<input type="email" name="email" placeholder="Enter Email" />
