+++ src/views/partials/front-header.ejs
@@ -20,11 +20,11 @@
 			>
 			<span
 				class="nav-link <% if (title === 'Afiye - Register Account') { %>active<% } %>"
-				><a href="/account/register">Register</a></span
+				><a href="/register">Register</a></span
 			>
 			<span
 				class="nav-link <% if (title === 'Afiye - Account Login') { %>active<% } %>"
-				><a href="/account/login">Login</a></span
+				><a href="/login">Login</a></span
 			>
 		</div>
 	</nav>
