+++ src/views/partials/user-header.ejs
@@ -19,7 +19,7 @@
 				class="nav-link <% if (title === 'Afiye - Account Settings') { %>active<% } %>"
 				><a href="/account/settings">Settings</a></span
 			>
-			<span class="nav-link"><a href="/account/logout">Logout</a></span>
+			<span class="nav-link"><a href="/logout">Logout</a></span>
 		</div>
 	</nav>
 </header>
