+++ src/views/register.ejs
@@ -4,7 +4,7 @@
 
 <%- include ('./partials/messages') %>
 
-<form action="/account/register" method="post">
+<form action="/register" method="post">
 	<div>
 		<label for="name">Name</label>
 		<input
